fix：修复押注限制值过大问题

+++ table/src/main/java/com/jjg/game/table/common/gamephase/BaseTableBetPhase.java
@@ -305,7 +305,7 @@ public abstract class BaseTableBetPhase<D extends TableGameDataVo> extends
             BetAreaCfg betAreaCfg = betAreaCfgMap.get(betAreaIdx);
             // 判断场上单区域的总数量是否达到上限
             // 配置的上限
-            int roomIdxMaxLimit = betAreaCfg.getTbUpperLimit() * betMax;
+            long roomIdxMaxLimit = (long) betAreaCfg.getTbUpperLimit() * betMax;
             // 当前房间的请求的下注区的总数
             long curIdxTotalBet = gameDataVo.getAreaTotalBet(betAreaIdx);
             if (curIdxTotalBet + betValue >= roomIdxMaxLimit) {
@@ -314,7 +314,7 @@ public abstract class BaseTableBetPhase<D extends TableGameDataVo> extends
                 return Code.BET_TO_LIMIT;
             }
             // 玩家区域上限
-            int playerIdxMaxLimit = betAreaCfg.getTbPlayerUpperLimit() * betMax;
+            long playerIdxMaxLimit = (long) betAreaCfg.getTbPlayerUpperLimit() * betMax;
             Map<Integer, List<Integer>> playerBetInfo = gameDataVo.getPlayerBetInfo(gamePlayer.getId());
             long playerBetTotal = 0;
             if (playerBetInfo != null) {
